+++ Software/Java/RuntimeQueue/src/main/java/com/mathworks/sparkbuilder/RuntimeQueueJanitor.java
@@ -1,0 +1,20 @@
+package com.mathworks.sparkbuilder;
+
+import java.util.TimerTask;
+
+public class RuntimeQueueJanitor extends TimerTask {
+
+    
+    @Override
+    public void run() {
+        RuntimeQueue.log("RuntimeQueueJanitor started.");
+
+        RuntimeQueue rtQueue = RuntimeQueue.getSingleton();
+
+        rtQueue.cleanupQueue(30L);
+
+        RuntimeQueue.log("RuntimeQueueJanitor finished.");
+
+
+    }
+}